+++ lib/mailchimp_api/collection_parsers/feedback.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module MailchimpAPI::CollectionParsers
+  class Feedback < Base
+    def element_key
+      'feedback'
+    end
+  end
+end